Move pipe_to_table from /gui to /outputs
