+++ rundmcmc/gui/__init__.py
@@ -1,2 +1,2 @@
-from .run import pipe_to_table, handle_scores_separately, run
+from .run import handle_scores_separately, run
 from .parse_config import read_basic_config
