+++ rundmcmc/gui/run.py
@@ -51,23 +51,6 @@ def handle_chain(chain, handlers):
         yield {key: handler(state) for key, handler in handlers.items()}
 
 
-def pipe_to_table(chain, handlers, display=True, number_to_display=10,
-                  number_to_bin=100):
-    table = ChainOutputTable()
-    display_interval = math.floor(len(chain) / number_to_display)
-    bin_interval = math.floor(len(chain) / number_to_bin)
-    counter = 0
-    for row in handle_chain(chain, handlers):
-        if counter % display_interval == 0:
-            if display:
-                print(f"Step {counter}")
-                print(row)
-        if counter % bin_interval == 0:
-            table.append(row)
-        counter += 1
-    return table
-
-
 def flips_to_dict(chain, handlers=None):
     hist = {0: chain.state.assignment}
     for state in chain:
