+++ rundmcmc/templates/template_main.py
@@ -35,9 +35,7 @@ from rundmcmc.validity import (L1_reciprocal_polsby_popper,
 from rundmcmc.scores import (efficiency_gap, mean_median,
                              mean_thirdian)
 
-from rundmcmc.run import pipe_to_table
-
-from rundmcmc.output import p_value_report
+from rundmcmc.output import p_value_report, pipe_to_table
 
 from vis_output import (hist_of_table_scores, trace_of_table_scores)
